removed irrelevant files and made minor adjustments
